[PAXEXAM-359] Mark org.ops4j.pax.exam.Inject deprecated
Although removal might be the better choice (as done for the
master/2.4.x) I prefer not to remove anything for a micro. I just set
the value deprecated to make sure nobody will use it and be paniced
about it not working as expected afterwards. As said, it's removed on
the master then.

Signed-off-by: Andreas Pieber <anpieber@gmail.com>