Allow a delay for a service to be present in the remote test container before installing the test probe
This modification can allow specific scenarios where a specialized
service is registered to indicate that the framework is ready for the
test. This can be especially useful for tests using blueprint or test
using bundles that load other bundles in the container without being
explicitely in the boot features defined by the pax exam test.