[PAXEXAM-740] Fix NoRouteToHostException RMI issue for Karaf
this appears e.g. on Fedora Linux on a laptop with two network adapters
(WiFi and wired Ethernet), if the hostname is not listed in /etc/hosts,
due to use of LocateRegistry -> InetAddress.getLocalHost(), but can also
occcur under other network configurations.

This change fixes the problem in pax-exam-container-karaf.  For
pax-exam-container-forked, changes in Pax Swissbox would also be
required.

see also
http://blog2.vorburger.ch/2017/04/java-local-rmi-using-jdks-inetaddress.html
for more background, and
https://bugs.opendaylight.org/show_bug.cgi?id=8176 for original error.