[PAXEXAM-740] Fix NoRouteToHostException RMI issue for forked
this appears e.g. on Fedora Linux on a laptop with two network adapters
(WiFi and wired Ethernet), if the hostname is not listed in /etc/hosts,
due to use of LocateRegistry -> InetAddress.getLocalHost(), but can also
occur under other network configurations.

This change fixes the problem in pax-exam-container-forked; which also
needed changes in Pax Swissbox, thus the uprade to a new version.

see also
http://blog2.vorburger.ch/2017/04/java-local-rmi-using-jdks-inetaddress.html