Fix Karaf[34]TestContainerITest
it was actually completely broken; as karafVersion() does not specify
which version to download, just adapt the start up; the frameworkUrl)(
needs a version() to choose between Karaf 3 & 4.. also can't be static.

Changed from ignore.ignoreLocalConsole().startRemoteShell() to
startLocalConsole().ignoreRemoteShell() instead, as it's handy to see
the Karaf banner on start up to recognize the version; and the remote
is not needed.

Also the versions were actually inavertedly inverted between *3/4Test* ;)