Remove dependency of reflection for TestNG
The PaxExam listener makes use of reflection
to disable configuration methods. 

This is NOT needed because the TestNG API has 
Provided setters via which these can be set.

The reason why this change is required is because
TestNG in its upcoming version is going to be 
Changing the data type of these two attributes 
viz “m_beforeTestMethods” and “m_afterTestMethods”